backend/rag-service/requirements.txt
# RAG Service Dependencies
# ============================================================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Vector Database
pymilvus>=2.4.0

# Embedding & NLP
sentence-transformers>=2.7.0
jieba>=0.42.1

# Authentication
python-jose[cryptography]>=3.3.0

# Settings
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
